basic namespace support, better support for rdf:Description
Namespace support is not very efficient, the whole URI is prepended to
name for comparaison, indexing should be used instead since there are
only a fairly small number of namespaces.

The parser no longer uses the global `objrdf::base_resource::nil' resource
it has been replaced by a member `placeholder', it is used to handle
that kind of RDF/XML:
...
<rdf:Description>
	<rdfs:type rdf:resource='...'/>
</rdf:Description>

Members have been ordered to match initialization list in constructor.