New allocator, resources are grouped by type
The goals of the new allocator are:
1. Group all the resources by type, each type has its own memory pool
dynamically allocated (array), those pools can written to a file in
raw binary possibly for later reuse, leading to persistence.
The grouping can be used to speed up SPARQL queries.
2. Pointers can be replaced by `pseudo' pointers: indexes into
the pool cell, they can be made a lot smaller than 8 bytes depending
on the type cardinality. For instance the number of rdfs::Classes is
known at compile time so a pointer to a rdfs::Class (like `rdf::type') should
only be ln2(n) bits. (pointers are templates that can be specialized)

More work is needed to deal with strings and property_array, strings
could be replaced by cstring of set size.