Generic pointer-like type with custom storage
Template parameters are used to set the storage back-end: free store or memory-mapped files, the pool size and the index size. Using the free store is not the same as using a regular pointer as the pseudo_ptr size depends mainly on the index. The shared_ptr implementation had to be slightly modified to handle pseudo_ptr.