fixed major bug with pool's id using hash
Each new pool was given a unique id based on a static counter, the
problem with that approach is that id depends on the execution path, if
in one path some class is not instantiated its pool wont be created and
all following pool's id will be offset by 1 causing havoc.
The solution adopted is to create the pool id with a hash of the class
qualified name eg: `rdfs_Class' (and information about the store and
the constness), only the least significant byte is used (risk of
collision) so it can be used as an index in a 255-cell pool.
More work needed to detect hash&0xff collision and hash&0xff==0.
When using a persistent store the pool id should be written to file so
it can be checked when loading the file.