Fixed major bug with persistent storage + improvements
There was a serious bug: when using persistent storage the file size was
ignored and only the first page was mapped causing major crashes.
The store has now access to the full pseudo_ptr implementation and can
get the maximum addressable space based on the pointer size.
It will also used in the future to facilitate mounting a file read-only.
The pointer size (pseudo_ptr::INDEX) is now a template parameter and
defaults to `unsigned short' that is maximum 2^32 objects.