simplified function table generation + transaction
There is hierarchy of classes that define all the function used by
properties, taking into account whether the property is literal or not,
whether the object is constant or not and so on.
First attempt at implementing transaction/version control, when a set
property is modified, copy is made of the object and then modification
is applied. It can be later decided to remove the copy.