fix definition of XML_Literal that caused crash
Storage information is now part of the class it defaults to
typedef persistent_store STORE;
in base_resource but can be overriden on a class basis (eg.
rdf::Property and rdfs::Class use free_store) there is still a STORE
parameter in the pseudo_ptr<> template but it is ignored.
The array<objrdf::superClassOf> is now populated right after the Class
is created in the function rdfs::Class::super().