change to pool allocation, management pools persisted
The two management pools: POOL_PTR and POOL_PTR_ARRAY are now saved
between runs of the program, this ensures that the same index is used
for the classes (using a different index would invalidate pointers).
When creating the different pools at the start of the program (or later
depending on the execution path) the management pools are look-up for a
matching class T (using hash function pseudo_ptr_pool<>::get_type_id<T>)
if it is found the new pool object is written over the old one (because
pointers, function pointers are no longer valid) if not found it is
created. After a few runs all the pools should be present in the
management pools.
This offer two major improvements:
.no risk of hash collision over 255, when the management pool is full it
will throw an exception (this will happen when we define more than 255
rdfs::Class)
.the pools are created at the same time as the rdfs::Class, as soon as
the rdfs::Class is created one can create instances