code cleanup, modification to const_instance_iterator
The iterator can now be used to return a pseudo property `self' that
points to the subject, that has applications in the SPARQL engine and
also in the future to return a different type of ID for each resource.
Added a new property rdfs:member, it won't be used as is but its
rdf:Property can be used to indicate that a property is a sub-class of a
container member. That helps creating forms to instantiate and edit
container-type classes.