major change to pseudo_ptr operation
Enforce standard C pointer conversion rule, it used to be possible to
cast a pseudo_ptr<const T> to pseudo_ptr<T> fixing that error broke code
in sparql_engine and rdf_xml_parser.
There is still a problem with information loss because of casting: a
const pointer might be converted to regular pointer and illegal changes
made to a resource.