new template parameter to pass property and class name
If NEW_NAME is defined the property and class names will be passed by a
structure with a single static member:
struct tmp{
	static char* name(){return "tmp";}
};
this fixes the multiple symbol problem and the special flag passed to
the linker is no longer needed. Once it seems stable the old code will
be removed.