new NATIVE flag
make it possible to use native pointer instead of fancy pointer, since
pools are no longer used we cannot iterate over a pool to get all
instances of a given class, and as a consequence we no longer have an
actual RDF document.
It is the user's responsibility to track all instances.
The advantage is faster compilation time (~50%) and smaller executables.